### 2019-05-17 #302 @mkearney
Submission: rtweet
Submitting Author: Michael W. Kearney (@mkearney)
Repository: https://github.com/mkearney/rtweet
Version submitted: v0.6.9
Version accepted: TBD

---

```
Package: rtweet
Version: 0.6.9
Title: Collecting Twitter Data
    person("Michael W.", "Kearney", ,
    email = "kearneymw@missouri.edu", role = c("aut", "cre"),
    ## add contributor template (middle name/initial optional)
    #person("First Middle", "Last", ,
    #email = "email@address.com", role = c("ctb"))
    )
Description: An implementation of calls designed to collect and
    organize Twitter data via Twitter's REST and stream Application
    Program Interfaces (API), which can be found at the following URL:
    <https://developer.twitter.com/en/docs>.
    jsonlite (>= 0.9.22),
    magrittr (>= 1.5.0),
    tibble (>= 1.3.4),
    progress,
    Rcpp,
    httpuv
    magick,
    openssl,
    webshot,
    covr,
    igraph
LazyData: yes
    Rcpp

```

**Data retrieval** because the package allows users to easily request and import data from Twitter's REST and stream APIs.

**Data munging** because significant work is done to convert JSON objects returned by Twitter's APIs into tabular data frames.

The target audience is researchers. The scientific applications span a range of topics. Here's is the relevant excerpt from [paper.md](https://github.com/mkearney/rtweet/blob/master/paper.md)
>Although ``rtweet`` provides some coverage to user context-behaviors (e.g.,
posting statuses, liking tweets, following users, etc.), the primary audience
for the package to date has been researchers. Accordingly, ``rtweet`` has been
featured in numerous popular press [e.g.,
@bajak2019democrats; @machlis2019r; @riley2019twitter] and academic publications
[e.g.,
@bossetta2018simulated; @bradley2019major; @buscema2018media;
@erlandsen2018twitter; @gitto2019brand; @kearney2019analyzing;
@kearney2018analyzing; @li2018sentiment; @lutkenhaus2019tailoring;
@lutkenhaus2019mapping; @molyneux2018media; @tsoi2018can; @unsihuay2018topic;
@valls2017urban; @wu2018finding].

To date no other package interfaces with both REST and stream APIs. The twitteR package is most similar, but it has entered a stage of deprecation (I've agreed to carry the torch, so to speak). So, not only does twitteR not reflect some recent changes to Twitter's API (most notably the introduction of 'extended tweet' mod–the new 280 character limit), but it lacks active maintenance thanks–in part–to rtweet filling the void.

https://github.com/ropensci/software-review/issues/193
