### 2019-05-21 #302 @mkearney
I tentatively listed @andrewheiss and @briatte as reviewers because they initially expressed their willingness to review on a [now closed] JOSS thread. I don't want to presume they are/will be the reviewers–I'm not completely familiar with the ropensci process, but I'd selection of reviewers isn't entirely up to me–but I figured reviewer suggestions/volunteers were worth knowing!
