### 2019-05-22 fdc566b
Fixing Inversion in tweet length allowed limit.
