### 2019-05-27 cde88dc
Correcting small docs error
