### 2019-05-28 #302 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @mkearney     !

Here's the output from [goodpractice](https://github.com/MangoTheCat/goodpractice/). If you haven't used `goodpractice` it's an R package that checks a number of things with another package - most of which we agree with and want authors to follow. You don't need to address these now, it's info for reviewers to use to get started.

- Regarding tests: all tests fail, because I don't have the auth set up. How do you suggest one gets set up to run your test suite? I clearly need a `twitter_tokens` file. How do I get that file? and where does it go?  Seems like a good thing to go in a `CONTRIBUTING.md` file or so - but maybe there's a good reason not to add that?

```r
── GP rtweet ─────

It is good practice to

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try make your lines
    shorter than 80 characters

    R/bearer_token.R:56:1
    R/bearer_token.R:80:1
    R/coords.R:119:1
    R/coords.R:131:1
    R/coords.R:132:1
    ... and 37 more lines

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data. Consider using vapply() instead.

    R/favorites.R:75:45
    R/post.R:326:12
    R/post.R:336:12
    R/stream.R:338:13
    R/utils.R:327:40
    ... and 1 more lines

  ✖ avoid the library() and require() functions, they change the global search path. If you need to use other packages, import them. If you need to
    load them explicitly, then consider loadNamespace() instead, or as a last resort, declare them as 'Depends' dependencies.

    R/utils.R:485:5

  ✖ fix this R CMD check NOTE: Note: found 113868 marked UTF-8 strings
  ✖ checking tests ...  SEE QUESTION ABOVE
```

Seeking reviewers now 🕐

---

Reviewers:

- xxx
- xxx
