### 2019-05-29 #302 @briatte
Here's version 1 of my review, with comments shown as list sub-items. I'll add the functionality tests soon.

## Package Review (draft version 1)

  - This step used to require Twitter API credentials, but does not anymore.
  - The package does not have a CONTRIBUTING file, but it does have the `URL` and `BugReports` fields in its DESCRIPTION. This seems sufficient to me.

>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).
>  - The `.bib` file is clean enough, but could be cleaned further (some DOI fields come as URLs, other as just DOIs).

Estimated hours spent reviewing: 0.5 (draft version 1)

---

Reviewed using R version 3.6.0 on x86_64-apple-darwin15.6.0 (64-bit).

Installed with

```r
devtools::install_github("mkearney/rtweet", dependencies = TRUE)
```
