### 2019-06-05 #302 @andrewheiss
>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

Estimated hours spent reviewing: ≈5 hours

---

### Documentation and vignettes

All vignettes worked great and were a perfect introduction to the package. It might be helpful to move up the vignette listing near the beginning of the README so that users can find them more easily.

Each function is generally very well documented. There are some sparsely documented functions like `bearer_token()` and `invalidate_bearer()` that seem to only be used internally—it might be useful to either add `@keywords internal` to the roxygen block or add additional documentation explaining what a bearer token is and including some examples.

It's really helpful that the US and the world are baked into `lookup_coords()` (e.g. `lookup_coords("usa")`) so users don't need an API key for that. This should probably be documented, though. It caught me by surprise that "usa" was working just fine and all other locations triggered a message—I only figured out why because I looked at the source code for `lookup_coords()`, and that has a huge all caps note about needing a Google Maps API key. (A future expansion could maybe include hardcoded boundaries for a bunch of other countries (perhaps make an internal lookup table of lat/long coordinates?) so the package is less US-centric and less dependent on another API.)

The **pkgdown** site is great. It might be helpful (and [recommended by rOpenSci](https://ropensci.github.io/dev_guide/building.html#function-grouping)) to add `@family` tags to roxygen documentation so that there are automatic groups of functions in the documentation and in the reference website.

### Installation and testing

Installation worked fine with `devtools::install_github("mkearney/rtweet", dependencies = TRUE)`

Testing doesn't work out of the box and it took me a while to figure out how to generate the `/tests/testthat/twitter_tokens` RDS file. When logged in through the embedded `rstats2twitter` app, running this created an auth token that the test suite can use:

```{r eval=FALSE}
saveRDS(get_token(), "tests/testthat/twitter_tokens")
```

It might be useful to explain that process somewhere in the documentation—not in the README though, since it's not something a typical user would do.

190 tests passed (super impressive coverage!!), and 5 tests failed:

- `direct_messages`: This is because the `rstats2twitter` app doesn't provide read-write DM access. It worked when I used my own auth token

- `get_my_timeline`: One test failed because `rtweet:::home_user() not equal to "kearneymw".` Mike's account is hardcoded in the tests and this test can't pass if someone is using a different account. I don't know what else could be done to test it though.

    ```
    test_get_my_timeline.R:6: failure: get_my_timeline
    rtweet:::home_user() not equal to "kearneymw".
    1/1 mismatches
    x[1]: "andrewheiss"
    y[1]: "kearneymw"
    ```

- `lookup_coords`: This returns an error that seems unrelated to token issues and I don't know while it's failing

    ```
    test_lookup_coords.R:36: error: lookup_coords returns coords data
    all arguments must be named
    1: do.call(Sys.setenv, ng) at /Users/andrew/Downloads/rtweet-master/tests/testthat/test_lookup_coords.R:36
    2: (function (...)
       {
           x <- list(...)
           nm <- names(x)
           if (is.null(nm) || "" %in% nm)
               stop("all arguments must be named")
           .Internal(Sys.setenv(nm, as.character(unlist(x))))
       })()
    3: stop("all arguments must be named")
    ──────────────────────────────────────────
    ```

- `tweet_shot`: This failed because I didn't have PhantomJS installed. The test worked after installing it. Running `tweet_shot()` interactively gives a note that it should be installed with `webshot::install_phantomjs()`, but this message is invisible when testing on a fresh system. PhantomJS isn't installed as a dependency when the package in installed, but that's because it's not an R package. idk what the best way to handle this is, though, since it's 16+ MB and not handled automatically with R ¯\\\_(ツ)\_/¯

- `test-test_username-r`: Like `get_my_timeline`, this failed because Mike's account is hardcoded in. Again, I don't know what the best way to handle this is.

    ```
    test_username.R:7: failure: test authenticating user name
    `sn` not equal to "kearneymw".
    1/1 mismatches
    x[1]: "andrewheiss"
    y[1]: "kearneymw"
    ```

### Token creation and authentication

It's so great that people can use this without creating their own Twitter apps, since that process has become long and arduous.

It might be helpful to more clearly explain the benefits of creating an app with tokens vs. just using the interactive browser-based token. Right now, the README says "You may still choose to do this [create an app] (gives you more stability and permissions)", and the auth vignette explains how to do it, but there isn't any explanation of why users might want to do it. Later in the README, under "Post actions," it explains that users would need their own app to post tweets and access DMs, but it could be helpful to have that information up above when describing authentication in general.

It might even be helpful to have a table of sorts showing pros/cons/features available when using a private app vs. the embedded `rstats2twitter` app, like this maybe?

| Task                         | rstats2twitter application   | Personal application   |
| :-----                       | :--------------------------: | :--------------------: |
| Work interactively           | ✅                            | ✅                      |
| Create reproducible research | ❌                            | ✅                      |
| Search tweets                | ✅                            | ✅                      |
| Stream tweets                | ✅                            | ✅                      |
| Get friends                  | ✅                            | ✅                      |
| Get timelines                | ✅                            | ✅                      |
| …                            | …                            | …                      |
| Post tweets                  | ❌                            | ✅                      |
| Run package tests            | ❌                            | ✅                      |

The page for creating apps looks slightly different from the screenshots, and the direction to go to apps.twitter.com is out of date—according to Twitter, that aspect of developer accounts is getting sunset. Instead, the URL is now https://developer.twitter.com/en/apps .

### Community

It would be helpful to have some [community guidelines](https://ropensci.github.io/dev_guide/collaboration.html#friendlyfiles) in the package too, such as a CONTRIBUTING.md file (perhaps modeled after something like [this](https://github.com/tidyverse/ggplot2/blob/master/CONTRIBUTING.md)) and a CONDUCT.md file (like [this](https://github.com/tidyverse/readr/blob/master/CONDUCT.md))

Creating an RStudio project `.Rproj` file could be helpful for encouraging package contributions, since it would enable RStudio's package building interface.

### rOpenSci-related issues

- [Following rOpenSci's package guidelines](https://ropensci.github.io/dev_guide/building.html#creating-metadata-for-your-package), it might be helpful to use the **codemetar** package to generate a `codemeta.json` file.
- Consider adding a repostatus.org badge too ([following rOpenSci's guidelines](https://ropensci.github.io/dev_guide/building.html#readme))
- Also add the rOpenSci peer review badge

### JOSS-related issues

There are a few issues with `paper.bib`:

- "Twitter as a tool of para-disploomacy" is misspelled
- Several of the cited articles that have actual DOIs are missing DOIs (like "How can we better use Twitter to find a person who got lost due to dementia?" is <https://dx.doi.org/10.1038/s41746-018-0017-5>)

# Kicking the tires with some use cases

Look at tweets

```
library(tidyverse)
library(rtweet)

# Check out my tweets
my_tweets <- get_timeline("andrewheiss", n = 500)

head(my_tweets)
#> # A tibble: 6 x 90
#>   user_id status_id created_at          screen_name text  source display_text_wi… reply_to_status…
#>   <chr>   <chr>     <dttm>              <chr>       <chr> <chr>             <dbl> <chr>
#> 1 167410… 11360038… 2019-06-04 20:16:30 andrewheiss @dva… Twitt…                0 113600361493217…
#> 2 167410… 11359881… 2019-06-04 19:14:09 andrewheiss I ME… Twitt…                6 113598739853218…
#> 3 167410… 11359873… 2019-06-04 19:11:16 andrewheiss Revi… Twitt…               94 NA
#> 4 167410… 11359848… 2019-06-04 19:01:06 andrewheiss "boo… Twitt…              171 113424196604718…
#> 5 167410… 11359781… 2019-06-04 18:34:42 andrewheiss oh m… Twitt…              140 NA
#> 6 167410… 11359750… 2019-06-04 18:22:13 andrewheiss As a… Twitt…              140 NA

ts_plot(my_tweets)
```

Geocoding with **sf**

```r
library(sf)
library(rnaturalearth)

searched_tweets <- search_tweets(
  q = "lang:en",
  geocode = lookup_coords("usa"), n = 10000
) %>%
  lat_lng() %>%
  drop_na(lng) %>%
  sf::st_as_sf(coords = c("lng", "lat"), crs = 4326)

us <- ne_states(country = "united states of america", returnclass = "sf") %>%
  filter(!(postal %in% c("HI", "AK", "PR")))

ggplot() +
  geom_sf(data = us) +
  geom_sf(data = searched_tweets) +
  coord_sf(crs = 102003, datum = NA) +  # Albers
  theme_void()
```

---

Magical. This is a great package—fantastic work!

---

Tested and reviewed using R 3.6.0 on x86_64-apple-darwin15.6.0 (64-bit).
