+++ DESCRIPTION
@@ -1,14 +1,14 @@
 Package: rtweet
 Type: Package
-Version: 0.6.9
+Version: 0.6.9.1
 Title: Collecting Twitter Data
 Authors@R: c(
     person("Michael W.", "Kearney", ,
     email = "kearneymw@missouri.edu", role = c("aut", "cre"),
     comment = c(ORCID = "0000-0002-0730-4694"))
     ## add contributor template (middle name/initial optional)
-    #person("First Middle", "Last", ,
-    #email = "email@address.com", role = c("ctb"))
+    person("James E", "Harris", ,
+    email = "jim@jeharrisllc.com", role = c("ctb"))
     )
 Description: An implementation of calls designed to collect and
     organize Twitter data via Twitter's REST and stream Application
