### 2019-06-25 #302 @briatte
Hi @sckott

I have updated my draft review, which can be considered final. My new comments are at the bottom of the review. In a nutshell, I hit the same issues as @andrewheiss did with the tests, and am asking for additional details to be included about how the package compares to `RTwitterAPI`.

Generally speaking, the package is awesome, thanks a lot for your work @mkearney!
