### 2019-06-26 mkearney/rtweet#340 @briatte
get_followers(retryonratelimit = TRUE) not working
### Problem

`get_followers` with `retryonratelimit = TRUE` does not actually retry anything: it downloads the first (by default) 5,000 followers, and exits.

The last part of my code, reproduced further down, does not perform a 'retry' but simply exits:

```
> x <- get_followers("fhollande", n = 5000, retryonratelimit = TRUE, token = t)
Downloading [====>------------------------------------]  13%75000 followers!
> nrow(x)
[1] 5000
```

### Expected behavior

Perhaps I misread the docs, but I would have expected the function to get at least 75,000 followers, then pause for 15 minutes, then continue, and so until finishing to collect the followers.

### Reproduce the problem

library(rtweet)

t <- create_token(
  app = "OAuth for @phnk (2019-06)",
  consumer_key = "...",
  consumer_secret = "...",
  access_token = "...",
  access_secret = "...",
  set_renv = TRUE
)

get_token() # checks out

# account with large amount of followers
x <- get_followers("fhollande", n = 5000, retryonratelimit = TRUE, token = t)

nrow(x) # 5000
```

### rtweet version

```
> packageVersion("rtweet")
[1] '0.6.9'
```

### Session info

```
> sessionInfo()
R version 3.6.0 (2019-04-26)
Platform: x86_64-apple-darwin15.6.0 (64-bit)
Running under: macOS High Sierra 10.13.6

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/3.6/Resources/lib/libRlapack.dylib

locale:
[1] en_GB.UTF-8/en_GB.UTF-8/en_GB.UTF-8/C/en_GB.UTF-8/en_GB.UTF-8

attached base packages:

other attached packages:
[1] rtweet_0.6.9

 [1] Rcpp_1.0.1        fansi_0.4.0       prettyunits_1.0.2 zeallot_0.1.0     utf8_1.1.4
 [6] assertthat_0.2.1  crayon_1.3.4      R6_2.4.0          backports_1.1.4   jsonlite_1.6
[11] magrittr_1.5      pillar_1.4.1      httr_1.4.0        cli_1.1.0         rlang_0.4.0
[16] progress_1.2.2    curl_3.3          rstudioapi_0.10   vctrs_0.1.0       tools_3.6.0
[21] hms_0.4.2         compiler_3.6.0    askpass_1.1       pkgconfig_2.0.2   openssl_1.4
[26] tibble_2.1.3
```

### Token

```
> rtweet::get_token()
<Token>
<oauth_endpoint>
 request:   https://api.twitter.com/oauth/request_token
 authorize: https://api.twitter.com/oauth/authenticate
 access:    https://api.twitter.com/oauth/access_token
<oauth_app> mytwitterapp
  key:    Jv6gEcY9548OWmKO2Z4WV8T0x
  secret: <hidden>
<credentials> oauth_token, oauth_token_secret
---
```

N.B. The token was recently created via https://developer.twitter.com, in replacement of an older (and, perhaps relatedly, dysfunctional) token that I held from when the Twitter API was called apps.twitter.com.
