### 2019-06-26 #302 @briatte
FYI, I've just posted an issue to `rtweet` that has to do with getting followers for accounts with over 5,000 followers: https://github.com/mkearney/rtweet/issues/340

Also, still related to followers, I subscribe to the suggestion made here to harmonize the output of `get_followers` and `get_friends`: https://github.com/mkearney/rtweet/issues/308
