### 2019-07-10 mkearney/rtweet#340 @Arf9999
Your code specifically sets the number of followers to 5000, so retryonratelimit does nothing.

If you want more followers, set n to a higher number.
