### 2019-08-19 000ac03
fix magrittr version
