+++ DESCRIPTION
@@ -19,7 +19,7 @@ Depends:
 Imports:
     httr (>= 1.3.0),
     jsonlite (>= 0.9.22),
-    magrittr (>= 1.5.0),
+    magrittr (>= 1.5),
     tibble (>= 1.3.4),
     utils,
     progress,
