### 2019-08-26 7fd6655
fix nchar validation in post_tweet
