+++ R/post.R
@@ -128,8 +128,16 @@ post_tweet <- function(status = "my first rtweet #rstats",
   on.exit(options(encoding = enc), add = TRUE)
   options(encoding = "UTF-8")
 
-  ## validate status text
-  if (all(is_tweet_length(status), !grepl("http", status))) {
+  ## validate status text – IF
+  ##   (Part 1) status text is > 280 characters
+  ##            ***AND***
+  ##   (Part 2) status text does not include hyperlink
+  ## ––––––––––logic:
+  ##   Twitter will shorten long URLs (so the characters in any supplied URL may
+  ##   not count 1:1 toward the 280 character limit); i'm not sure when and how
+  ##   this works (we'd need to know exactly *when* and *to what extent* URLs
+  ##   get shorted), so this is an inexact solution
+  if (all(!is_tweet_length(status), !grepl("https?://\\S+", status))) {
     stop("cannot exceed 280 characters.", call. = FALSE)
   }
   if (length(status) > 1) {
