### 2019-08-26 63c1c3b
Augment documentation for bearer_token()
