### 2019-08-26 eb31d75
Augment documentation for bearer_token()
