### 2019-08-26 5ab72da
Fix pkg require checks in tests
