+++ README.Rmd
@@ -63,13 +63,45 @@ All you need is a **Twitter account** (user name and password) and you can be up
 
 Simply send a request to Twitter's API (with a function like `search_tweets()`, `get_timeline()`, `get_followers()`, `get_favorites()`, etc.) during an interactive session of R, authorize the embedded **`rstats2twitter`** app (approve the browser popup), and your token will be created and saved/stored (for future sessions) for you!
 
-## API authorization
+### API authorization
 
 All users must be authorized to interact with Twitter's APIs. To become authorized, simply use a function like `search_tweets()`, `get_timeline()`, `get_followers()`, or `get_favorites()` in an interactive session an authorize via web browser popup on behalf of your Twitter account!
 
 **It is no longer necessary to obtain a developer account and create your own Twitter application to use Twitter's API.** You may still choose to do this (gives you more stability and permissions), but **{rtweet}** should work out of the box assuming (a) you are working in an interactive/live session of R and (b) you have installed the [**{httpuv}**](https://github.com/rstudio/httpuv) package.
 
 + If you still want to apply for a developer account and create your own application, see the `auth` vignette (or the API authorization section below) for additional instructions: [https://rtweet.info/articles/auth.html](https://rtweet.info/articles/auth.html).
+
+### Vignettes
+
+[Obtaining and using Twitter API tokens](https://rtweet.info/articles/auth.html)
+
+```{r}
+## quick overview of rtweet functions
+vignette("auth", package = "rtweet")
+```
+
+
+[Quick overview of rtweet package](https://rtweet.info/articles/intro.html)
+
+```{r}
+## quick overview of rtweet functions
+vignette("intro", package = "rtweet")
+```
+
+[Live streaming tweets data](https://rtweet.info/articles/stream.html)
+
+```{r}
+## working with the stream
+vignette("stream", package = "rtweet")
+```
+
+[Troubleshooting common rtweet problems](https://rtweet.info/articles/FAQ.html)
+
+```{r}
+## working with the stream
+vignette("FAQ", package = "rtweet")
+```
+
 
 ### Package features
 
@@ -267,36 +299,6 @@ Twitter app
 - To create your own Twitter app, follow the instructions in the
   authorization vignette on [obtaining and using access tokens](https://rtweet.info/articles/auth.html)
 
-#### Vignettes
-
-[Obtaining and using Twitter API tokens](https://rtweet.info/articles/auth.html)
-
-```{r}
-## quick overview of rtweet functions
-vignette("auth", package = "rtweet")
-```
-
-
-[Quick overview of rtweet package](https://rtweet.info/articles/intro.html)
-
-```{r}
-## quick overview of rtweet functions
-vignette("intro", package = "rtweet")
-```
-
-[Live streaming tweets data](https://rtweet.info/articles/stream.html)
-
-```{r}
-## working with the stream
-vignette("stream", package = "rtweet")
-```
-
-[Troubleshooting common rtweet problems](https://rtweet.info/articles/FAQ.html)
-
-```{r}
-## working with the stream
-vignette("FAQ", package = "rtweet")
-```
 
 ## Contact
 
