### 2019-08-26 0562c1c
Add reviewers and update pkg docs
