+++ NEWS.md
@@ -1,3 +1,7 @@
+- Added paper.md as part of ROpenSci submission.
+- Added contributing template.
+- Added explanation of requirements and usage to `bearer_token()` docs.
+
 # rtweet 0.6.9
 - Better tweet-validating in streaming data–interrupted statuses/broken lines 
   are now returned
