+++ vignettes/reviewer-reply.Rmd
@@ -1,0 +1,104 @@
+---
+title: "Reviewer Reply"
+output: github_document
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
+```
+
+## andrewwheiss
+
+### Documentation and vignettes
+
+> Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.
+
+* [x] Added @andrewwheis as reviewer role to DESCRIPTION
+
+> All vignettes worked great and were a perfect introduction to the package. It might be helpful to move up the vignette listing near the beginning of the README so that users can find them more easily.
+
+* [x] Moved vignettes section up to above key functions (and just after install/API authorization)
+ 
+> Each function is generally very well documented. There are some sparsely documented functions like bearer_token() and invalidate_bearer() that seem to only be used internally—it might be useful to either add @keywords internal to the roxygen block or add additional documentation explaining what a bearer token is and including some examples.
+
+* [x] Added `@ keywords internal` to `invalidate_bearer()`
+* [x] Improved name/description, added more information about requirements and usage, and several examples to `bearer_token()`
+
+> It's really helpful that the US and the world are baked into lookup_coords() (e.g. lookup_coords("usa")) so users don't need an API key for that. This should probably be documented, though. It caught me by surprise that "usa" was working just fine and all other locations triggered a message—I only figured out why because I looked at the source code for lookup_coords(), and that has a huge all caps note about needing a Google Maps API key. (A future expansion could maybe include hardcoded boundaries for a bunch of other countries (perhaps make an internal lookup table of lat/long coordinates?) so the package is less US-centric and less dependent on another API.)
+
+* [ ] Document US/world coordinates in `lookup_coords()`
+* [ ] Add other (less US-centric) defaults to `lookup_coords()`
+
+> The pkgdown site is great. It might be helpful (and recommended by rOpenSci) to add @family tags to roxygen documentation so that there are automatic groups of functions in the documentation and in the reference website.
+
+* [ ] Add `@family` tags for reference/function documentation grouping
+
+### Installation and testing
+
+
+> Testing doesn't work out of the box and it took me a while to figure out how to generate the /tests/testthat/twitter_tokens RDS file. When logged in through the embedded rstats2twitter app, running this created an auth token that the test suite can use...It might be useful to explain that process somewhere in the documentation—not in the README though, since it's not something a typical user would do.
+
+* [ ] Added description of configuration/setup of test token
+
+> (FAILED) direct_messages: This is because the rstats2twitter app doesn't provide read-write DM access. It worked when I used my own auth token
+
+* [x] Testing now checks for appropriate permissions.
+* [x] As for the limited permissions, this was a concsious decision on my part. I limited write and DM-related permissions in tokens generated via rstats2twitter because I don't want to encourage or be associated with automated/spam/TOS-violating behaviors.
+
+> (FAILED) get_my_timeline: One test failed because rtweet:::home_user() not equal to "kearneymw". Mike's account is hardcoded in the tests and this test can't pass if someone is using a different account. I don't know what else could be done to test it though.
+
+* [x] Agreed. This may have been possible back when I had a dedicated rtweet Twitter account, but that got shut down by Twitter (I learned you can't use multiple accounts/tokens to do the same thing the hard way), so I'm hesitant/careful not to further hurt my standing with them.
+
+> (FAILED) lookup_coords: This returns an error that seems unrelated to token issues and I don't know while it's failing
+
[…]
