+++ DESCRIPTION
@@ -4,8 +4,14 @@ Version: 0.6.9
 Title: Collecting Twitter Data
 Authors@R: c(
     person("Michael W.", "Kearney", ,
-    email = "kearneymw@missouri.edu", role = c("aut", "cre"),
-    comment = c(ORCID = "0000-0002-0730-4694"))
+      email = "kearneymw@missouri.edu", role = c("aut", "cre"),
+      comment = c(ORCID = "0000-0002-0730-4694")),
+    person("Andrew Heiss", "Heiss", ,
+      role = c("rev"),
+      comment = c(ORCID = "0000-0002-3948-3914")),
+    person("Andrew Heiss", "Heiss", ,
+      role = c("rev"),
+      comment = c(ORCID = "0000-0002-3948-3914"))
     ## add contributor template (middle name/initial optional)
     #person("First Middle", "Last", ,
     #email = "email@address.com", role = c("ctb"))
