### 2019-09-04 df4d33f
add citycoords
