+++ R/coords.R
@@ -18,6 +18,18 @@
 #' @param apikey A valid Google Maps API key. If NULL, `lookup_coords()` will
 #'   look for a relevant API key stored as an environment variable (e.g.,
 #'   `GOOGLE_MAPS_KEY`).
+#' @details Since Google Maps implemented stricter API requirements, sending 
+#'   requests to Google's API isn't very convenient. To enable basic uses
+#'   without requiring a Google Maps API key, a number of the major cities
+#'   throughout the word and the following two larger locations are 
+#'   baked into this function: 'world' and 'usa'. If 'world' is supplied then
+#'   a bounding box of maximum latitutde/longitude values, i.e.,
+#'   \code{c(-180, -90, 180, 90)}, and a center point \code{c(0, 0)} are 
+#'   returned. If 'usa' is supplied then estimates of the United States' 
+#'   bounding box and mid-point are returned. To specify a city, provide the
+#'   city name followed by a space and then the US state abbreviation or 
+#'   country name. To see a list of all included cities, enter
+#'   \code{rtweet:::citycoords} in the R console to see coordinates data.
 #' @param ... Additional arguments passed as parameters in the HTTP
 #'   request
 #' @return Object of class coords.
@@ -69,6 +81,18 @@ lookup_coords <- function(address, components = NULL, apikey = NULL, ...) {
     point <- c(
       lat = 0,
       lng = 0
+    )
+  } else if (gsub("\\,\\s?", " ", tolower(address)) %in% citycoords$city) {
+    address <- gsub("\\,\\s?", " ", tolower(address))
+    i <- match(address, citycoords$city)
+    point <- c(
+      lat = citycoords$lat[i],
+      lng = citycoords$lng[i])
+    boxp <- c(
+      sw.lng = point[2] - .05,
+      sw.lat = point[1] - .05,
+      ne.lng = point[2] + .05,
+      ne.lat = point[1] + .05
     )
   } else {
     ## encode address
