+++ tests/testthat/test_lookup_coords.R
@@ -3,7 +3,12 @@ context("lookup_coords")
 
 test_that("lookup_coords returns coords data", {
   skip_on_cran()
-
+  
+  kcmo <- lookup_coords("kansas city, mo")
+  expect_gt(cor(kcmo$point, c(39.0997, 94.5786)), 0.9)
+  tor <- lookup_coords("toronto canada")
+  expect_gt(cor(tor$point, c(43.6532, 79.3832)), 0.9)
+  
   x <- lookup_coords("usa")
   expect_equal(is.list(x), TRUE)
   expect_named(x)
