### 2019-09-05 42dbad1
Update auth vignette and add code of conduct
