+++ vignettes/auth.Rmd
@@ -19,19 +19,26 @@ This vignette covers how to obtain and use Twitter API access tokens for use in 
 
 ## Creating a Twitter App
 
-- To create a Twitter app, navigate to [apps.twitter.com](https://apps.twitter.com/) and create a new app by providing a `Name`, `Description`, and `Website` of your choosing (example screenshot provided below).
+- To create a Twitter app, you will first need to apply for a developer account. This process along with a detailed explanation can be found at [developer.twitter.com](https://developer.twitter.com).
 
-- **Important** In the `Callback URL` field, make sure to enter the following: `http://127.0.0.1:1410`
-
-- Check yes if you agree and then click "Create your Twitter application".
-
-<p align="center">
-<img src="files/creating.png" alt="creating">
-</p>
+- Once you have acquired a developer account, navigate to [developer.twitter.com/en/apps](https://developer.twitter.com/en/apps/), click the blue button that says, `Create a New App`, and then complete the form with the following fields:
+   + `App Name`: What your app will be called
+   + `Application Description`: How your app will be described to its users
+   <p align="center"><img src="files/create-app-1.png" alt="create-app-1"></p>
+   + `Website URLs`: Website associated with app–I recommend using the URL to your Twitter profile
+   + `Callback URLs`: ***IMPORTANT*** enter exactly the following: `http://127.0.0.1:1410`
+   <p align="center"><img src="files/create-app-2.png" alt="create-app-2"></p>
+   + `Tell us how this app will be used`: Be clear and honest
+   <p align="center"><img src="files/create-app-3.png" alt="create-app-3"></p>
+- When you've completed the required form fields, click the blue `Create` button at the bottom
+- Read through and indicate whether you accept the developer terms
+   <p align="center"><img src="files/create-app-4.png" alt="create-app-4"></p>
+- And you're done!
+   <p align="center"><img src="files/create-app-5.png" alt="create-app-5"></p>
 
 ## Authorization methods
 
-Users can create their personal access token in two different ways. Each method is outlined below.
+Users can create their personal Twitter token in two different ways. Each method is outlined below.
 
 ### 1. Browser-based authentication
 
@@ -44,82 +51,76 @@ if (!requireNamespace("httpuv", quietly = TRUE)) {
 }
 ```
 
-- Click the tab labeled `Keys and Access Tokens` to retrieve your keys.
+- Navigate to [developer.twitter.com/en/apps](https://developer.twitter.com/en/apps) and select your Twitter app
+- Click the tab labeled `Keys and tokens` to retrieve your keys.
+- Locate the `Consumer API keys` (aka "API Secret").
 
-<p align="center">
-<img src="files/created.png" alt="created">
-</p>
-
-- In the `Keys and Access Tokens` tab, locate the values `Consumer Key` (aka "API Key") and `Consumer Secret` (aka "API Secret").
-
-<p align="center">
-<img src="files/keys.png" alt="keys">
-</p>
+<p align="center"><img src="files/create-app-6.png" alt="create-app-6"></p>
 
 - Copy and paste the two keys (along with the name of your app) into an R script file and pass them along to `create_token()`.
 
 ```{r}
-## autheticate via web browser
+## load rtweet
+library(rtweet)
+
+## store api keys (these are fake example values; replace with your own keys)
+api_key <- "afYS4vbIlPAj096E60c4W1fiK"
+api_secret_key <- "bI91kqnqFoNCrZFbsjAWHD4gJ91LQAhdCJXCj3yscfuULtNkuu"
+
[…]
