+++ vignettes/reviewer-reply.Rmd
@@ -7,7 +7,66 @@ output: github_document
 knitr::opts_chunk$set(echo = TRUE)
 ```
 
-## andrewwheiss
+## @briatte
+
+> Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.
+
+* [x] Added @briatte as reviewer role to DESCRIPTION
+
+> Tested with both the "out of the box" access to the Twitter API and personal OAuth credentials via create_token -- both worked as expected, although I ran into exactly the same list of issues as @andrewheiss when testing the package with testthat::auto_test_package().
+
+* [x] This is addressed in my reply to @andrewwheis below
+
+> Typo at https://rtweet.info/articles/auth.html -- "autheticate via web browser".
+
+* [x] Fixed typo (`autheticate` to `authetincate`) in auth vignette.
+
+> Could the thread about how rwteet compares to other packages also document how the package compares to RTwitterAPI?
+
+* [x] Adding a package comparison table to README. Note: The RTwitterAPI package is outdated and very bare bones, but while it does technically offer a way to hit most endpoints it doesn't do much to automate that process, parse the response, or document endpoints besides a 'get friends' example. And, thus, I've used the red question mark on the relevant GET API endpoints for RTwitterAPI.
+
+| Task                   | rtweet | twitteR | streamR | RTwitterAPI |
+|------------------------------|--------|---------|---------|-------------|
+| Available on CRAN            |   ✅   |    ✅   |    ✅   |      ❌     |
+| Updated since 2016           |   ✅   |    ❌   |    ✅   |      ❌     |
+| Non-'developer' access       |   ✅   |    ❌   |    ❌   |      ❌     |
+| Extended tweets (280 chars)  |   ✅   |    ❌   |    ✅   |      ❌     |
+| Parses JSON data             |   ✅   |    ✅   |    ✅   |      ❌     |
+| Converts t❌data frames      |   ✅   |    ✅   |    ✅   |      ❌     |
+| Automated pagination         |   ✅   |    ❌   |    ❌   |      ❌     |
+| Search tweets                |   ✅   |    ✅   |    ❌   |      ❓     |
+| Search users                 |   ✅   |    ❌   |    ❌   |      ❓     |
+| Stream sample                |   ✅   |    ❌   |    ✅   |      ❌     |
+| Stream keywords              |   ✅   |    ❌   |    ✅   |      ❌     |
+| Stream users                 |   ✅   |    ❌   |    ✅   |      ❌     |
+| Get friends                  |   ✅   |    ✅   |    ❌   |      ✅     |
+| Get timelines                |   ✅   |    ✅   |    ❌   |      ❓     |
+| Get mentions                 |   ✅   |    ✅   |    ❌   |      ❓     |
+| Get favorites                |   ✅   |    ✅   |    ❌   |      ❓     |
+| Get trends                   |   ✅   |    ✅   |    ❌   |      ❓     |
+| Get list members             |   ✅   |    ❌   |    ❌   |      ❓     |
+| Get list memberships         |   ✅   |    ❌   |    ❌   |      ❓     |
+| Get list statuses            |   ✅   |    ❌   |    ❌   |      ❓     |
+| Get list subscribers         |   ✅   |    ❌   |    ❌   |      ❓     |
+| Get list subscriptions       |   ✅   |    ❌   |    ❌   |      ❓     |
+| Get list users               |   ✅   |    ❌   |    ❌   |      ❓     |
+| Lookup collections           |   ✅   |    ❌   |    ❌   |      ❓     |
+| Lookup friendships           |   ✅   |    ✅   |    ❌   |      ❓     |
+| Lookup statuses              |   ✅   |    ✅   |    ❌   |      ❓     |
[…]
