+++ DESCRIPTION
@@ -6,12 +6,11 @@ Authors@R: c(
     person("Michael W.", "Kearney", ,
       email = "kearneymw@missouri.edu", role = c("aut", "cre"),
       comment = c(ORCID = "0000-0002-0730-4694")),
-    person("Andrew Heiss", "Heiss", ,
+    person("Andrew", "Heiss", ,
       role = c("rev"),
       comment = c(ORCID = "0000-0002-3948-3914")),
-    person("Andrew Heiss", "Heiss", ,
-      role = c("rev"),
-      comment = c(ORCID = "0000-0002-3948-3914"))
+    person("Francois", "Briatte", ,
+      role = c("rev"))
     ## add contributor template (middle name/initial optional)
     #person("First Middle", "Last", ,
     #email = "email@address.com", role = c("ctb"))
