### 2019-09-05 4aa13e4
Add codemeta and rOpenSci badge
