### 2019-09-05 09f6c06
Add/fix badges
