+++ README.md
@@ -10,7 +10,10 @@ status](https://www.r-pkg.org/badges/version/rtweet)](https://cran.r-project.org
 [![Coverage
 Status](https://codecov.io/gh/mkearney/rtweet/branch/master/graph/badge.svg)](https://codecov.io/gh/mkearney/rtweet?branch=master)
 [![DOI](https://zenodo.org/badge/64161359.svg)](https://zenodo.org/badge/latestdoi/64161359)
-[![](https://badges.ropensci.org/%3Cissue_id%3E_status.svg)](https://github.com/ropensci/software-review/issues/%3Cissue_id%3E)
+[![](https://badges.ropensci.org/1454_status.svg)](https://github.com/ropensci/software-review/issues/1454)
+[![Project Status: Active – The project has reached a stable, usable
+state and is being actively
+developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 ![Downloads](https://cranlogs.r-pkg.org/badges/rtweet)
 ![Downloads](https://cranlogs.r-pkg.org/badges/grand-total/rtweet)
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
