+++ paper.bib
@@ -41,7 +41,8 @@
   title = {A Simulated Cyberattack on {T}witter: Assessing Partisan Vulnerability to Spear Phishing and Disinformation ahead of the 2018 {US} Midterm Elections},
   author = {Bossetta, Michael},
   journal = {First Monday},
-  year = {2018}
+  year = {2018},
+  url = {https://ui.adsabs.harvard.edu/link_gateway/2018arXiv181105900B/arxiv:1811.05900}
 }
 
 @article{bradley2019major,
@@ -64,14 +65,15 @@
 
 
 @article{erlandsen2018twitter,
-  title = {Twitter as a tool of para-disploomacy: An exploratory cohort study based on {C}atalonia (2013-2017)},
+  title = {Twitter as a tool of para-diplomacy: An exploratory cohort study based on {C}atalonia (2013-2017)},
   author = {Erlandsen, Matthias},
   journal = {Revista Chilena de Relaciones Internacionales},
   volume = {2},
   issue = {1},
   year = {2018},
   pages = {211-231},
-  url = {https://rchri.cl/wp-content/uploads/2018/04/211-231.pdf}
+  url = {https://rchri.cl/wp-content/uploads/2018/04/211-231.pdf},
+  doi = {10.5281/zenodo.1188429.}
 }
 
 
@@ -131,7 +133,8 @@
   journal={Vaccine: X},
   pages={100019},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {https://doi.org/10.1016/j.jvacx.2019.100019}
 }
 
 
@@ -153,7 +156,8 @@
   number = {1},
   pages = {14},
   year = {2018},
-  publisher = {Nature Publishing Group}
+  publisher = {Nature Publishing Group},
+  doi = {https://dx.doi.org/10.1038/s41746-018-0017-5}
 }
 
 @article{unsihuay2018topic,
