+++ README.Rmd
@@ -19,13 +19,58 @@ knitr::opts_chunk$set(
 [![CRAN status](https://www.r-pkg.org/badges/version/rtweet)](https://cran.r-project.org/package=rtweet)
 [![Coverage Status](https://codecov.io/gh/mkearney/rtweet/branch/master/graph/badge.svg)](https://codecov.io/gh/mkearney/rtweet?branch=master)
 [![DOI](https://zenodo.org/badge/64161359.svg)](https://zenodo.org/badge/latestdoi/64161359)
-[![](https://badges.ropensci.org/1454_status.svg)](https://github.com/ropensci/software-review/issues/1454)
+[![](https://badges.ropensci.org/302_status.svg)](https://github.com/ropensci/software-review/issues/302)
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 ![Downloads](https://cranlogs.r-pkg.org/badges/rtweet)
 ![Downloads](https://cranlogs.r-pkg.org/badges/grand-total/rtweet)
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 
 R client for accessing Twitter's REST and stream APIs. Check out the [rtweet package documentation website](https://rtweet.info).
+
+### Package Functionality
+
+There are several R packages for interacting with Twitter's APIs. See how {rtweet} compares to these others in the chart below.
+
+| Task                         | [rtweet](https://github.com/mkearney/rtweet) | [twitteR](https://github.com/geoffjentry/twitteR) | [twitteR](https://github.com/pablobarbera/streamR) | [RTwitterAPI](https://github.com/joyofdata/RTwitterAPI) |
+|------------------------------|--------|---------|---------|-------------|
+| Available on CRAN            |   ✅   |    ✅   |    ✅   |      ❌     |
+| Updated since 2016           |   ✅   |    ❌   |    ✅   |      ❌     |
+| Non-'developer' access       |   ✅   |    ❌   |    ❌   |      ❌     |
+| Extended tweets (280 chars)  |   ✅   |    ❌   |    ✅   |      ❌     |
+| Parses JSON data             |   ✅   |    ✅   |    ✅   |      ❌     |
+| Converts to data frames      |   ✅   |    ✅   |    ✅   |      ❌     |
+| Automated pagination         |   ✅   |    ❌   |    ❌   |      ❌     |
+| Search tweets                |   ✅   |    ✅   |    ❌   |      ❓     |
+| Search users                 |   ✅   |    ❌   |    ❌   |      ❓     |
+| Stream sample                |   ✅   |    ❌   |    ✅   |      ❌     |
+| Stream keywords              |   ✅   |    ❌   |    ✅   |      ❌     |
+| Stream users                 |   ✅   |    ❌   |    ✅   |      ❌     |
+| Get friends                  |   ✅   |    ✅   |    ❌   |      ✅     |
+| Get timelines                |   ✅   |    ✅   |    ❌   |      ❓     |
+| Get mentions                 |   ✅   |    ✅   |    ❌   |      ❓     |
+| Get favorites                |   ✅   |    ✅   |    ❌   |      ❓     |
+| Get trends                   |   ✅   |    ✅   |    ❌   |      ❓     |
+| Get list members             |   ✅   |    ❌   |    ❌   |      ❓     |
+| Get list memberships         |   ✅   |    ❌   |    ❌   |      ❓     |
+| Get list statuses            |   ✅   |    ❌   |    ❌   |      ❓     |
[…]
