### 2019-09-05 c99c7fe
Fix pkg names in comp table
