### 2019-09-06 #302 @andrewheiss
Other than a duplicate "Premium 30 day" row in the big comparison table, this addresses all my concerns and I think it's good to go ✅
