### 2019-09-06 2555119
Remove duplicate pkg comp entry as pointed out here: https://github.com/ropensci/software-review/issues/302#issuecomment-528973489
