+++ vignettes/reviewer-reply.Rmd
@@ -74,8 +74,7 @@ Thank,
 | Post mute                    |   ✅   |    ❌   |    ❌   |      ❌     |
 | Premium 30 day               |   ✅   |    ❌   |    ❌   |      ❌     |
 | Premium full archive         |   ✅   |    ❌   |    ❌   |      ❌     |
-| Premium 30 day               |   ✅   |    ❌   |    ❌   |      ❌     |
-| Run package tests            |   ✅   |    ❌   |    ❌   |      ❌     |
+| Pun package tests            |   ✅   |    ❌   |    ❌   |      ❌     |
 
 ## @andrewwheiss
 
