### 2019-10-09 #302 @sckott
@briatte Are you happy with the changes made?
