### 2019-10-16 #302 @sckott
@mkearney let's move on. I'll take a final look and get back to you ASAP
