### 2019-10-17 #302 @sckott
looks good, just a single comment: `^codemeta\.json$` should be in .Rbuildignore

Approved! Thanks @mkearney for submitting and @briatte @andrewheiss for your reviews!

- [ ] for the `pkgdown` site, we usually have people point to `https://docs.ropensci.org/package_name` but it looks like you have your own domain name https://rtweet.info/ , and I imagine you'd prefer it live there
- [ ] Add a mention of the review in `DESCRIPTION` via [`rodev::add_ro_desc()`](https://docs.ropensci.org/rodev/reference/add_ro_desc.html).

For JOSS:

- [ ] Submit to JOSS at <https://joss.theoj.org/papers/new>, using the rOpenSci GitHub repo URL. When a JOSS "PRE REVIEW" issue is generated for your paper, add the comment: `This package has been reviewed by rOpenSci: https://LINK.TO/THE/REVIEW/ISSUE, @ropensci/editors`
