### 2019-10-19 #302 @mkearney
@sckott I’m not currently seeing an invite (and I wasn’t authorized to transfer to ropensci). Where would I find this?
