### 2019-10-20 #302 @briatte
> @briatte Are you happy with the changes made?

Yes! With apologies for not signalling it earlier.
