### 2019-10-21 59bc590
fix truncated tweet text in lists_statuses
