### 2019-10-21 ccf454e
change Github links to point to ropensci
