+++ paper.bib
@@ -2,7 +2,7 @@
   title = {Github issue #1: What differences with twitte{R}?},
   author = {Michael W Kearney},
   year = {2016},
-  url = {https://github.com/mkearney/rtweet/issues/1#issuecomment-492753003},
+  url = {https://github.com/ropensci/rtweet/issues/1#issuecomment-492753003},
   organization = {Github}
 }
 
