+++ README.md
@@ -4,11 +4,11 @@
 # rtweet <img src="man/figures/logo.png" width="160px" align="right" />
 
 [![Build
-Status](https://travis-ci.org/mkearney/rtweet.svg?branch=master)](https://travis-ci.org/mkearney/rtweet)
+Status](https://travis-ci.org/ropensci/rtweet.svg?branch=master)](https://travis-ci.org/ropensci/rtweet)
 [![CRAN
 status](https://www.r-pkg.org/badges/version/rtweet)](https://cran.r-project.org/package=rtweet)
 [![Coverage
-Status](https://codecov.io/gh/mkearney/rtweet/branch/master/graph/badge.svg)](https://codecov.io/gh/mkearney/rtweet?branch=master)
+Status](https://codecov.io/gh/ropensci/rtweet/branch/master/graph/badge.svg)](https://codecov.io/gh/ropensci/rtweet?branch=master)
 [![DOI](https://zenodo.org/badge/64161359.svg)](https://zenodo.org/badge/latestdoi/64161359)
 [![](https://badges.ropensci.org/302_status.svg)](https://github.com/ropensci/software-review/issues/302)
 [![Project Status: Active – The project has reached a stable, usable
@@ -27,7 +27,7 @@ There are several R packages for interacting with Twitter’s APIs. See
 how {rtweet} compares to these others in the chart
 below.
 
-| Task                        | [rtweet](https://github.com/mkearney/rtweet) | [twitteR](https://github.com/geoffjentry/twitteR) | [streamR](https://github.com/pablobarbera/streamR) | [RTwitterAPI](https://github.com/joyofdata/RTwitterAPI) |
+| Task                        | [rtweet](https://github.com/ropensci/rtweet) | [twitteR](https://github.com/geoffjentry/twitteR) | [streamR](https://github.com/pablobarbera/streamR) | [RTwitterAPI](https://github.com/joyofdata/RTwitterAPI) |
 | --------------------------- | -------------------------------------------- | ------------------------------------------------- | -------------------------------------------------- | ------------------------------------------------------- |
 | Available on CRAN           | ✅                                            | ✅                                                 | ✅                                                  | ❌                                                       |
 | Updated since 2016          | ✅                                            | ❌                                                 | ✅                                                  | ❌                                                       |
@@ -94,7 +94,7 @@ if (!requireNamespace("remotes", quietly = TRUE)) {
 }
 
 ## install dev version of rtweet from github
-remotes::install_github("mkearney/rtweet")
+remotes::install_github("ropensci/rtweet")
 
 ## load rtweet package
 library(rtweet)
@@ -384,6 +384,9 @@ sf <- get_trends("san francisco")
 Communicating with Twitter’s APIs relies on an internet connection,
 which can sometimes be inconsistent. With that said, if you encounter an
 obvious bug for which there is not already an active
-[issue](https://github.com/mkearney/rtweet/issues), please [create a new
-issue](https://github.com/mkearney/rtweet/issues/new) with all code used
-(preferably a reproducible example) on Github.
+[issue](https://github.com/ropensci/rtweet/issues), please [create a new
+issue](https://github.com/ropensci/rtweet/issues/new) with all code used
+(preferably a reproducible example) on
+Github.
+
+[![ropensci\_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
