### 2019-10-21 5921bad
Fix pandoc rendering
