### 2019-10-22 dd9b013
Fix refs and dois in paper.bib
