+++ paper.bib
@@ -30,7 +30,7 @@
   year = {2018},
   note = {R package version 0.6.7},
   url = {https://cran.r-project.org/package=rtweet},
-  doi = {https://doi.org/10.5281/zenodo.2528481}
+  doi = {10.5281/zenodo.2528481}
 }
 
 
@@ -38,11 +38,15 @@
 %% ACADEMIC JOURNAL ARTICLES USING RTWEET
 
 @article{bossetta2018simulated,
-  title = {A Simulated Cyberattack on {T}witter: Assessing Partisan Vulnerability to Spear Phishing and Disinformation ahead of the 2018 {US} Midterm Elections},
-  author = {Bossetta, Michael},
-  journal = {First Monday},
-  year = {2018},
-  url = {https://ui.adsabs.harvard.edu/link_gateway/2018arXiv181105900B/arxiv:1811.05900}
+	author = {Michael Bossetta},
+	title = {A simulated cyberattack on {T}witter: Assessing partisan vulnerability to spear phishing and disinformation ahead of the 2018 U.S. midterm elections},
+	journal = {First Monday},
+	volume = {23},
+	number = {12},
+	year = {2018},
+	issn = {13960466},
+	doi = {10.5210/fm.v23i12.9540},
+	url = {https://firstmonday.org/ojs/index.php/fm/article/view/9540}
 }
 
 @article{bradley2019major,
@@ -52,7 +56,7 @@
   pages={1--20},
   year={2019},
   publisher={Taylor \& Francis},
-  doi={https://doi.org/10.1080/14459795.2019.1606927}
+  doi={10.1080/14459795.2019.1606927}
 }
 
 @article{buscema2018media,
@@ -85,7 +89,7 @@
   pages = {153 - 163},
   year = {2019},
   issn = {0969-6997},
-  doi = {https://doi.org/10.1016/j.jairtraman.2019.01.010},
+  doi = {10.1016/j.jairtraman.2019.01.010},
   url = {http://www.sciencedirect.com/science/article/pii/S0969699718303144}
 }
 
@@ -94,7 +98,7 @@
   year   = {2019},
   journal = {New Media \& Society},
   title   = {Analyzing change in network polarization},
-  url     = {https://doi.org/10.1177/1461444818822813},
+  url     = {10.1177/1461444818822813},
   note    = {[Online First]},
   doi     = {10.1177/1461444818822813}
 }
@@ -108,13 +112,16 @@
   publisher = {ABC-CLIO}
 }
 
-@article{li2018sentiment,
-  title = {Sentiment-based prediction of alternative cryptocurrency price fluctuations using gradient boosting tree model},
-  author = {Li, Tianyu Ray and Chamrajnagar, Anup S and Fong, Xander R and Rizik, Nicholas R and Fu, Feng},
-  journal = {arXiv preprint arXiv:1805.00558},
-  year = {2018}
+@article{li2019sentiment,
+  title={Sentiment-based prediction of alternative cryptocurrency price fluctuations using gradient boosting tree model},
+  author={Li, Tianyu Ray and Chamrajnagar, Anup and Fong, Xander and Rizik, Nicholas and Fu, Feng},
+  journal={Frontiers in Physics},
+  volume={7},
+  pages={98},
+  year={2019},
+  publisher={Frontiers},
+  doi={10.3389/fphy.2019.00098}
 }
-
 
 @article{lutkenhaus2019tailoring,
   title = {Tailoring in the digital era: Stimulating dialogues on health topics in collaboration with social media influencers},
@@ -134,7 +141,7 @@
   pages={100019},
   year={2019},
[…]
