+++ paper.bib
@@ -77,7 +77,7 @@
   year = {2018},
   pages = {211-231},
   url = {https://rchri.cl/wp-content/uploads/2018/04/211-231.pdf},
-  doi = {10.5281/zenodo.1188429.}
+  doi = {10.5281/zenodo.1188429}
 }
 
 
@@ -151,7 +151,7 @@
   journal = {New Media \& Society},
   pages = {1-20},
   year = {2018},
-  doi = {10.1177/F1461444818809392}
+  doi = {10.1177/1461444818809392}
 }
 
 
@@ -164,7 +164,7 @@
   pages = {14},
   year = {2018},
   publisher = {Nature Publishing Group},
-  doi = {https://dx.doi.org/10.1038/s41746-018-0017-5}
+  doi = {10.1038/s41746-018-0017-5}
 }
 
 @article{unsihuay2018topic,
