### 2019-10-22 #302 @stefaniebutland
@mkearney When would you like to publish the post? Tues Nov 12 is earliest date available, and that would mean submitting a draft via pull request by Tues Nov 5. Later dates are also open.

Details https://github.com/ropensci/software-review/issues/302#issuecomment-529040066
