### 2019-10-23 #302 @jeroen
Docs are now live on https://docs.ropensci.org/rtweet/
