### 2019-10-24 #302 @mkearney
@stefaniebutland Nov 5th/12th works for me!
