### 2019-11-06 #302 @stefaniebutland
@mkearney Will you submit a draft post soon?
Technical and editorial guidelines: https://github.com/ropensci/roweb2#contributing-a-blog-post
